SameDiff reorganization - first pass
